Cleanup using IntelliJ
@shale-stripe convinced me that this is better than using Eclipse

+++ src/main/java/com/stripe/net/Webhook.java
@@ -212,7 +212,7 @@ public final class Webhook {
 		 * @return the timestamp as a long.
 		 */
 		public static long getTimeNow() {
-			long time = (long)(System.currentTimeMillis() / 1000L);
+			long time = System.currentTimeMillis() / 1000L;
 			return time;
 		}
 	}
